+++ examples/experimental/dagster-dlift/kitchen-sink/dlift_kitchen_sink_tests/conftest.py
@@ -1,10 +1,10 @@
 import pytest
-from dagster_dlift.cloud_instance import DbtCloudInstance
+from dagster_dlift.client import DbtCloudClient
 from dlift_kitchen_sink.instance import get_environment_id, get_instance
 
 
 @pytest.fixture
-def instance() -> DbtCloudInstance:
+def instance() -> DbtCloudClient:
     return get_instance()
