+++ examples/experimental/dagster-dlift/kitchen-sink/dlift_kitchen_sink_tests/test_instance_from_env.py
@@ -1,11 +1,11 @@
-from dagster_dlift.cloud_instance import DbtCloudInstance
+from dagster_dlift.client import DbtCloudClient
 from dlift_kitchen_sink.instance import get_environment_id, get_instance
 
 
 def test_cloud_instance() -> None:
     """Test that we can create a DbtCloudInstance and verify connections successfully."""
     instance = get_instance()
-    assert isinstance(instance, DbtCloudInstance)
+    assert isinstance(instance, DbtCloudClient)
 
     instance.verify_connections()
